// contracts/typescript/models/SelectedDay.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HomeOfficeReservation API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface SelectedDay
 */
export interface SelectedDay {
    /**
     *
     */
    dayId: string;
    /**
     *
     */
    expectedVersion: number | null;
}

/**
 * Check if a given object implements the SelectedDay interface.
 */
export function instanceOfSelectedDay(value: object): value is SelectedDay {
    if (!('dayId' in value) || value['dayId'] === undefined) return false;
    if (!('expectedVersion' in value) || value['expectedVersion'] === undefined) return false;
    return true;
}

export function SelectedDayFromJSON(json: any): SelectedDay {
    return SelectedDayFromJSONTyped(json, false);
}

export function SelectedDayFromJSONTyped(json: any, ignoreDiscriminator: boolean): SelectedDay {
    if (json == null) {
        return json;
    }
    return {

        'dayId': json['dayId'],
        'expectedVersion': json['expectedVersion'],
    };
}

export function SelectedDayToJSON(json: any): SelectedDay {
    return SelectedDayToJSONTyped(json, false);
}

export function SelectedDayToJSONTyped(value?: SelectedDay | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'dayId': value['dayId'],
        'expectedVersion': value['expectedVersion'],
    };
}